Generate server from webpack to make use of loaders
