+++ config/webpack.config.prod.js
@@ -41,10 +41,70 @@ if (env['process.env.NODE_ENV'] !== '"production"') {
   throw new Error('Production builds must have NODE_ENV=production.');
 }
 
+const loaders = [
+  // Process JS with Babel.
+  {
+    test: /\.(js|jsx)$/,
+    include: [paths.appSrc, paths.appScripts],
+    loader: 'babel',
+
+  },
+  // The notation here is somewhat confusing.
+  // "postcss" loader applies autoprefixer to our CSS.
+  // "css" loader resolves paths in CSS and adds assets as dependencies.
+  // "style" loader normally turns CSS into JS modules injecting <style>,
+  // but unlike in development configuration, we do something different.
+  // `ExtractTextPlugin` first applies the "postcss" and "css" loaders
+  // (second argument), then grabs the result CSS and puts it into a
+  // separate file in our build process. This way we actually ship
+  // a single CSS file in production instead of JS code injecting <style>
+  // tags. If you use code splitting, however, any async bundles will still
+  // use the "style" loader inside the async code so CSS from them won't be
+  // in the main CSS file.
+  {
+    test: /\.css$/,
+    // "?-autoprefixer" disables autoprefixer in css-loader itself:
+    // https://github.com/webpack/css-loader/issues/281
+    // We already have it thanks to postcss. We only pass this flag in
+    // production because "css" loader only enables autoprefixer-powered
+    // removal of unnecessary prefixes when Uglify plugin is enabled.
+    // Webpack 1.x uses Uglify plugin as a signal to minify *all* the assets
+    // including CSS. This is confusing and will be removed in Webpack 2:
+    // https://github.com/webpack/webpack/issues/283
+    loader: ExtractTextPlugin.extract('style', 'css?-autoprefixer!postcss')
+    // Note: this won't work without `new ExtractTextPlugin()` in `plugins`.
+  },
+  // JSON is not enabled by default in Webpack but both Node and Browserify
+  // allow it implicitly so we also enable it.
+  {
+    test: /\.json$/,
+    loader: 'json'
+  },
+  // "file" loader makes sure those assets end up in the `build` folder.
+  // When you `import` an asset, you get its filename.
+  {
+    test: /\.(ico|jpg|jpeg|png|gif|eot|otf|webp|svg|ttf|woff|woff2)(\?.*)?$/,
+    loader: 'file',
+    query: {
+      name: 'images/[name].[hash:8].[ext]'
+    }
+  },
+  // "url" loader works just like "file" loader but it also embeds
+  // assets smaller than specified size as data URLs to avoid requests.
+  {
+    test: /\.(mp4|webm|wav|mp3|m4a|aac|oga)(\?.*)?$/,
+    loader: 'url',
+    query: {
+      limit: 10000,
+      name: 'images/[name].[hash:8].[ext]'
+    }
+  }
+]
+
 // This is the production configuration.
 // It compiles slowly and is focused on producing a fast and minimal bundle.
 // The development configuration is different and lives in a separate file.
-module.exports = {
+module.exports = [{
   // Don't attempt to continue if there are any errors.
   bail: true,
   // We generate sourcemaps in production. This is slow but gives good results.
@@ -94,65 +154,7 @@ module.exports = {
         include: paths.appSrc
       }
     ],
-    loaders: [
-      // Process JS with Babel.
-      {
-        test: /\.(js|jsx)$/,
-        include: paths.appSrc,
-        loader: 'babel',
-
-      },
-      // The notation here is somewhat confusing.
-      // "postcss" loader applies autoprefixer to our CSS.
-      // "css" loader resolves paths in CSS and adds assets as dependencies.
-      // "style" loader normally turns CSS into JS modules injecting <style>,
-      // but unlike in development configuration, we do something different.
-      // `ExtractTextPlugin` first applies the "postcss" and "css" loaders
-      // (second argument), then grabs the result CSS and puts it into a
-      // separate file in our build process. This way we actually ship
-      // a single CSS file in production instead of JS code injecting <style>
-      // tags. If you use code splitting, however, any async bundles will still
-      // use the "style" loader inside the async code so CSS from them won't be
-      // in the main CSS file.
-      {
-        test: /\.css$/,
-        // "?-autoprefixer" disables autoprefixer in css-loader itself:
-        // https://github.com/webpack/css-loader/issues/281
-        // We already have it thanks to postcss. We only pass this flag in
-        // production because "css" loader only enables autoprefixer-powered
-        // removal of unnecessary prefixes when Uglify plugin is enabled.
-        // Webpack 1.x uses Uglify plugin as a signal to minify *all* the assets
-        // including CSS. This is confusing and will be removed in Webpack 2:
-        // https://github.com/webpack/webpack/issues/283
-        loader: ExtractTextPlugin.extract('style', 'css?-autoprefixer!postcss')
-        // Note: this won't work without `new ExtractTextPlugin()` in `plugins`.
-      },
-      // JSON is not enabled by default in Webpack but both Node and Browserify
-      // allow it implicitly so we also enable it.
-      {
-        test: /\.json$/,
-        loader: 'json'
-      },
-      // "file" loader makes sure those assets end up in the `build` folder.
-      // When you `import` an asset, you get its filename.
-      {
-        test: /\.(ico|jpg|jpeg|png|gif|eot|otf|webp|svg|ttf|woff|woff2)(\?.*)?$/,
-        loader: 'file',
-        query: {
-          name: 'images/[name].[hash:8].[ext]'
-        }
-      },
-      // "url" loader works just like "file" loader but it also embeds
-      // assets smaller than specified size as data URLs to avoid requests.
-      {
-        test: /\.(mp4|webm|wav|mp3|m4a|aac|oga)(\?.*)?$/,
-        loader: 'url',
-        query: {
-          limit: 10000,
-          name: 'images/[name].[hash:8].[ext]'
-        }
-      }
-    ]
+    loaders: loaders
   },
 
   // We use PostCSS for autoprefixing only.
@@ -226,4 +228,40 @@ module.exports = {
     net: 'empty',
     tls: 'empty'
   }
-};
+}, {
+  name: 'server-side render',
+  entry: './scripts/render',
+  target: 'node',
+  output: {
+    path: './build',
+    filename: 'render.webpack.js',
+    publicPath: publicPath,
+    libraryTarget: "commonjs2"
+  },
+  externals: /^[a-z\-0-9]+$/,
+  module: {
+    loaders: loaders
+  },
+  plugins: [
+    new webpack.DefinePlugin(env),
+    // This helps ensure the builds are consistent if source hasn't changed:
+    new webpack.optimize.OccurrenceOrderPlugin(),
+    // Try to dedupe duplicated modules, if any:
+    new webpack.optimize.DedupePlugin(),
+    // Minify the code.
+    new webpack.optimize.UglifyJsPlugin({
+      compress: {
+        screw_ie8: true, // React doesn't support IE8
+        warnings: false
+      },
+      mangle: {
+        screw_ie8: true
+      },
+      output: {
+        comments: false,
+        screw_ie8: true
+      }
+    }),
+    new ExtractTextPlugin('[name].css')
+  ]
+}];
